ZBuffering and Transparency

+++ triebWerk/include/CMesh.h
@@ -17,7 +17,7 @@ namespace triebWerk
 	public:
 		CMesh::SVertex* m_pVertices;
 		ID3D11Buffer* m_pVertexBuffer;
-		unsigned int m_VertexCount;
+		size_t m_VertexCount;
 
 	public:
 		CMesh();
