+++ triebWerk/src/CPhysicWorld.cpp
@@ -196,7 +196,7 @@ void triebWerk::CPhysicWorld::CheckCollisionEvents()
     {
         ICollider* collider = m_DynamicCollider[i];
 
-        for (int j = collider->m_CollisionEvents.size() - 1; j >= 0; --j)
+        for (size_t j = collider->m_CollisionEvents.size() - 1; j >= 0; --j)
         {
             CCollisionEvent& collEvent = collider->m_CollisionEvents[j];
